feat: separate login hooks
