+++ src/app/login/components/LoginForm.jsx
@@ -1,34 +1,9 @@
 "use client";
 
-import { useRouter } from "next/navigation";
-import React, { useState } from "react";
-import Cookies from "js-cookie";
+import { useLogin } from "../hooks/useLogin";
 
 export const LoginForm = () => {
-  const router = useRouter();
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-
-  async function handleSubmitLogin(e) {
-    e.preventDefault();
-
-    const res = await fetch("https://eventmakers-api.fly.dev/auth/login", {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({
-        email,
-        password,
-      }),
-    });
-    const { payload, token } = await res.json();
-
-    localStorage.setItem("user", JSON.stringify(payload));
-
-    Cookies.set("token", token);
-    router.push("/dashboard");
-  }
+  const { setEmail, setPassword, handleSubmitLogin } = useLogin();
 
   return (
     <div className="flex w-full h-full">
